refactor: :recycle: Changed u32 types in gpu temp to f32

+++ src/gpu.rs
@@ -4,25 +4,25 @@ use nvml_wrapper::{
 };
 use nvml_wrapper::NVML;
 
-pub fn get_gpu_avg_temp()  -> Result<u32, NvmlError> {
-    let temps: Vec<u32> = get_gpu_temps().unwrap();
-    let cnt: u32 = temps.len() as u32;
-    let mut nums: u32 = 0;
+pub fn get_gpu_avg_temp()  -> Result<f32, NvmlError> {
+    let temps: Vec<f32> = get_gpu_temps().unwrap();
+    let cnt: f32 = temps.len() as f32;
+    let mut nums: f32 = 0f32;
     for element in temps.iter() {
         nums += *element;
     }
-    Ok((nums / cnt) as u32)
+    Ok(nums / cnt)
 }
 
-fn get_gpu_temps() -> Result<Vec<u32>, NvmlError> {
+fn get_gpu_temps() -> Result<Vec<f32>, NvmlError> {
     let nvml: NVML = NVML::init().unwrap();
     let device_count: u32 = nvml.device_count().unwrap();
-    let mut temperatures: Vec<u32> = vec![];
+    let mut temperatures: Vec<f32> = vec![];
 
     for i in 0..device_count {
         let device: Device = nvml.device_by_index(i).unwrap();
         let temp: u32 = read_gpu_temp(&device).unwrap();
-        temperatures.push(temp);
+        temperatures.push(temp as f32);
     }
     Ok(temperatures)
